ScanCode(Test): Move parser-related code to the new ScanCodeResultParser(Test)
While at it, fix the package for ScanCodeResultParserTest.
