+++ scanner/src/main/kotlin/scanners/ScanCode.kt
@@ -23,7 +23,6 @@ import java.io.File
 import java.io.IOException
 import java.net.HttpURLConnection
 import java.time.Instant
-import java.util.regex.Pattern
 
 import kotlin.math.max
 
@@ -34,7 +33,6 @@ import okio.sink
 
 import org.apache.logging.log4j.Level
 
-import org.ossreviewtoolkit.model.OrtIssue
 import org.ossreviewtoolkit.model.Provenance
 import org.ossreviewtoolkit.model.ScanResult
 import org.ossreviewtoolkit.model.ScannerDetails
@@ -80,7 +78,7 @@ class ScanCode(
         const val SCANNER_NAME = "ScanCode"
 
         private const val OUTPUT_FORMAT = "json-pp"
-        private const val TIMEOUT = 300
+        internal const val TIMEOUT = 300
 
         /**
          * Configuration options that are relevant for [getConfiguration] because they change the result file.
@@ -117,83 +115,6 @@ class ScanCode(
             "--$OUTPUT_FORMAT"
         } else {
             "--output-$OUTPUT_FORMAT"
-        }
-
-        // Note: The "(File: ...)" part in the patterns below is actually added by our own getRawResult() function.
-        private val UNKNOWN_ERROR_REGEX = Pattern.compile(
-            "(ERROR: for scanner: (?<scanner>\\w+):\n)?" +
-                    "ERROR: Unknown error:\n.+\n(?<error>\\w+Error)(:|\n)(?<message>.*) \\(File: (?<file>.+)\\)",
-            Pattern.DOTALL
-        )
-
-        private val TIMEOUT_ERROR_REGEX = Pattern.compile(
-            "(ERROR: for scanner: (?<scanner>\\w+):\n)?" +
-                    "ERROR: Processing interrupted: timeout after (?<timeout>\\d+) seconds. \\(File: (?<file>.+)\\)"
-        )
-
-        /**
-         * Map messages about unknown issues to a more compact form. Return true if solely memory errors occurred,
-         * return false otherwise.
-         */
-        internal fun mapUnknownIssues(issues: MutableList<OrtIssue>): Boolean {
-            if (issues.isEmpty()) {
-                return false
-            }
-
-            var onlyMemoryErrors = true
-
-            val mappedIssues = issues.map { fullError ->
-                UNKNOWN_ERROR_REGEX.matcher(fullError.message).let { matcher ->
-                    if (matcher.matches()) {
-                        val file = matcher.group("file")
-                        val error = matcher.group("error")
-                        if (error == "MemoryError") {
-                            fullError.copy(message = "ERROR: MemoryError while scanning file '$file'.")
-                        } else {
-                            onlyMemoryErrors = false
-                            val message = matcher.group("message").trim()
-                            fullError.copy(message = "ERROR: $error while scanning file '$file' ($message).")
-                        }
-                    } else {
-                        onlyMemoryErrors = false
-                        fullError
-                    }
-                }
-            }
-
-            issues.clear()
-            issues += mappedIssues.distinctBy { it.message }
-
-            return onlyMemoryErrors
-        }
-
-        /**
-         * Map messages about timeout errors to a more compact form. Return true if solely timeout errors occurred,
-         * return false otherwise.
-         */
-        internal fun mapTimeoutErrors(issues: MutableList<OrtIssue>): Boolean {
-            if (issues.isEmpty()) {
-                return false
-            }
-
-            var onlyTimeoutErrors = true
-
-            val mappedIssues = issues.map { fullError ->
-                TIMEOUT_ERROR_REGEX.matcher(fullError.message).let { matcher ->
-                    if (matcher.matches() && matcher.group("timeout") == TIMEOUT.toString()) {
-                        val file = matcher.group("file")
-                        fullError.copy(message = "ERROR: Timeout after $TIMEOUT seconds while scanning file '$file'.")
-                    } else {
-                        onlyTimeoutErrors = false
-                        fullError
-                    }
-                }
-            }
-
-            issues.clear()
-            issues += mappedIssues.distinctBy { it.message }
-
-            return onlyTimeoutErrors
         }
     }
